ENH: reorganize spectrochempy-nmr with readers/ and extern/ packages
