+++ plugins/spectrochempy-nmr/src/spectrochempy_nmr/readers/__init__.py
@@ -1,0 +1,3 @@
+# ======================================================================================
+# NMR reader modules for SpectroChemPy.
+# ======================================================================================
