+++ plugins/spectrochempy-nmr/src/spectrochempy_nmr/readers/read_agilent.py
@@ -33,7 +33,7 @@ from spectrochempy.core.units import ur
 from spectrochempy.utils._logging import warning_
 from spectrochempy.utils.meta import Meta
 
-from .nmrglue import read_varian
+from spectrochempy_nmr.extern.nmrglue import read_varian
 
 # ======================================================================================
 # Valid Agilent parameter names for metadata extraction
