+++ plugins/spectrochempy-nmr/tests/test_read_agilent.py
@@ -42,7 +42,7 @@ class TestAgilentNMRGlue:
     """Tests for the vendored nmrglue Agilent reading functions."""
 
     def test_read_varian_procpar(self):
-        from spectrochempy_nmr.nmrglue import read_varian_procpar
+        from spectrochempy_nmr.extern.nmrglue import read_varian_procpar
 
         pdic = read_varian_procpar(str(AGILENT_DIR / "agilent_1d" / "procpar"))
         assert isinstance(pdic, dict)
@@ -53,16 +53,16 @@ class TestAgilentNMRGlue:
         assert pdic["sw"]["values"] == ["50000"]
 
     def test_find_varian_shape_1d(self):
-        from spectrochempy_nmr.nmrglue import find_varian_shape
-        from spectrochempy_nmr.nmrglue import read_varian_procpar
+        from spectrochempy_nmr.extern.nmrglue import find_varian_shape
+        from spectrochempy_nmr.extern.nmrglue import read_varian_procpar
 
         pdic = read_varian_procpar(str(AGILENT_DIR / "agilent_1d" / "procpar"))
         shape = find_varian_shape(pdic)
         assert shape == (1500,)
 
     def test_find_varian_shape_2d(self):
-        from spectrochempy_nmr.nmrglue import find_varian_shape
-        from spectrochempy_nmr.nmrglue import read_varian_procpar
+        from spectrochempy_nmr.extern.nmrglue import find_varian_shape
+        from spectrochempy_nmr.extern.nmrglue import read_varian_procpar
 
         pdic = read_varian_procpar(str(AGILENT_DIR / "agilent_2d" / "procpar"))
         shape = find_varian_shape(pdic)
@@ -70,7 +70,7 @@ class TestAgilentNMRGlue:
         assert shape[0] == 332  # ni * 2 phase values
 
     def test_read_varian_fid_1d(self):
-        from spectrochempy_nmr.nmrglue import read_varian
+        from spectrochempy_nmr.extern.nmrglue import read_varian
 
         dic, data = read_varian(
             str(AGILENT_DIR / "agilent_1d"),
@@ -82,7 +82,7 @@ class TestAgilentNMRGlue:
         assert np.issubdtype(data.dtype, np.complexfloating)
 
     def test_read_varian_fid_2d(self):
-        from spectrochempy_nmr.nmrglue import read_varian
+        from spectrochempy_nmr.extern.nmrglue import read_varian
 
         dic, data = read_varian(
             str(AGILENT_DIR / "agilent_2d"),
@@ -94,9 +94,9 @@ class TestAgilentNMRGlue:
         assert np.issubdtype(data.dtype, np.complexfloating)
 
     def test_find_varian_torder(self):
-        from spectrochempy_nmr.nmrglue import find_varian_shape
-        from spectrochempy_nmr.nmrglue import find_varian_torder
-        from spectrochempy_nmr.nmrglue import read_varian_procpar
+        from spectrochempy_nmr.extern.nmrglue import find_varian_shape
+        from spectrochempy_nmr.extern.nmrglue import find_varian_torder
+        from spectrochempy_nmr.extern.nmrglue import read_varian_procpar
 
         pdic = read_varian_procpar(str(AGILENT_DIR / "agilent_2d" / "procpar"))
         shape = find_varian_shape(pdic)
@@ -104,7 +104,7 @@ class TestAgilentNMRGlue:
         assert torder in ("f", "r", "o")
 
     def test_procpar_key_count(self):
-        from spectrochempy_nmr.nmrglue import read_varian_procpar
+        from spectrochempy_nmr.extern.nmrglue import read_varian_procpar
 
         pdic = read_varian_procpar(str(AGILENT_DIR / "agilent_1d" / "procpar"))
         assert len(pdic) > 100  # typical procpar has 100+ parameters
